Split job&snapshot coordination in MasterContext
This is supposed to be refactor only, even though it's hard to compare...
