+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/execution/init/EdgeDef.java
@@ -20,7 +20,7 @@ import com.hazelcast.jet.config.EdgeConfig;
 import com.hazelcast.jet.core.Edge;
 import com.hazelcast.jet.core.Edge.RoutingPolicy;
 import com.hazelcast.jet.core.Partitioner;
-import com.hazelcast.jet.impl.MasterContext;
+import com.hazelcast.jet.impl.MasterJobContext;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
@@ -99,7 +99,7 @@ public class EdgeDef implements IdentifiedDataSerializable {
     }
 
     boolean isSnapshotRestoreEdge() {
-        return priority == MasterContext.SNAPSHOT_RESTORE_EDGE_PRIORITY;
+        return priority == MasterJobContext.SNAPSHOT_RESTORE_EDGE_PRIORITY;
     }
 
     boolean isDistributed() {
