+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/execution/init/VertexDef.java
@@ -17,7 +17,7 @@
 package com.hazelcast.jet.impl.execution.init;
 
 import com.hazelcast.jet.core.ProcessorSupplier;
-import com.hazelcast.jet.impl.MasterContext;
+import com.hazelcast.jet.impl.MasterJobContext;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
@@ -81,7 +81,7 @@ public class VertexDef implements IdentifiedDataSerializable {
     }
 
     boolean isSnapshotVertex() {
-        return name.startsWith(MasterContext.SNAPSHOT_VERTEX_PREFIX);
+        return name.startsWith(MasterJobContext.SNAPSHOT_VERTEX_PREFIX);
     }
 
     /**
